Move missed processor
